reorganize code.
